requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic-settings==2.2.1
httpx
python-dotenv

langgraph
langchain
langchain-groq
langchain-openai

transformers
torch
accelerate
sentencepiece
safetensors
peft
bitsandbytes
sqlalchemy
asyncpg
psycopg2-binary
alembic
loguru==0.7.2
email-validator
passlib[bcrypt]
python-jose[cryptography]
